Adds user owned resource and emulation project scaffolding

+++ eaasi-web-api/src/controllers/base/BaseUserOwnedCrudController.ts
@@ -1,65 +1,26 @@
-import ICrudController from '@/controllers/interfaces/ICrudController';
 import { build_400_response, build_404_response, build_500_response } from '@/utils/error-helpers';
 import HttpResponseCode from '@/classes/HttpResponseCode';
-import { areAllValidIntegerParams } from '@/utils/validators';
-import { Request, Response } from 'express';
+import { Response } from 'express';
 import { Result } from 'express-validator';
-import CrudService from 'src/services/base/CrudService';
-import BaseController from './BaseController';
-import { Model } from 'sequelize-typescript';
-
+import { EaasiUserOwnedModel } from '@/data_access/models/app/base-models/EaasiIUserOwnedModel';
+import BaseCrudController from './BaseCrudController';
+import { IAuthorizedRequest } from '@/types/auth/Auth';
+import { EaasiRoles } from '@/types/auth/User';
 
 /**
- * Base class for Controllers that handle CRUD logic
+ * Base class for Controllers that handle user-owned CRUD logic
  */
-export default class BaseCrudController<T extends Model> extends BaseController implements ICrudController {
-
-	//eslint-disable-next-line @typescript-eslint/no-explicit-any
-	protected service: CrudService<T>;
-
-	//eslint-disable-next-line @typescript-eslint/no-explicit-any
-	constructor(crudService: CrudService<T>) {
-		super();
-		this.service = crudService;
-	}
-
-	/**
-	 * Returns all object instances in paginated form.
-	 * @param req request
-	 * @param res response
-	 */
-	async getAll(req: Request, res: Response) {
-		let query = this._getQueryFromParams(req);
-
-		// todo: investigate more robust query string validation, add sortCol validation
-		if (!areAllValidIntegerParams([query.limit, query.page])) {
-			return res
-				.status(HttpResponseCode.BAD_REQUEST)
-				.send(build_400_response(JSON.stringify(req.query)));
-		}
-
-		let response = await this.service.getAll(req.query);
-		const err: Error = response.error instanceof Error ? response.error : new Error(response.error);
-		if (response.hasError) {
-			return res
-				.status(HttpResponseCode.SERVER_ERROR)
-				.send(build_500_response(err));
-		}
-
-		return res.send(response.result);
-	}
+export default class BaseUserOwnedCrudController<T extends EaasiUserOwnedModel> extends BaseCrudController<T> {
 
 	/**
 	 * Gets a resource by ID
 	 * @param req request
 	 * @param res response
 	 */
-	async get(req: Request, res: Response) {
-		// @ts-ignore
-		const id = req.params.id;
+	async get(req: IAuthorizedRequest, res: Response) {
+		const id = Number(req.params['id']);
 
-		// @ts-ignore
-		if (req.params.id == null) {
+		if (req.params['id'] == null) {
 			return res
 				.status(HttpResponseCode.BAD_REQUEST)
 				// @ts-ignore
@@ -80,6 +41,10 @@ export default class BaseCrudController<T extends Model> extends BaseController 
 				.send(build_404_response(req.originalUrl));
 		}
 
+		if (response.result.userId !== req.user.id) {
+			return res.status(HttpResponseCode.UNAUTHORIZED).send();
+		}
+
 		return res.status(HttpResponseCode.OK).send(response.result);
 	}
 
@@ -88,13 +53,13 @@ export default class BaseCrudController<T extends Model> extends BaseController 
 	 * @param req request
 	 * @param res response
 	 */
-	async create(req: Request, res: Response) {
+	async create(req: IAuthorizedRequest, res: Response) {
 		const newObject = req.body;
 
 		if (newObject == null) {
 			return res
 				.status(HttpResponseCode.BAD_REQUEST)
-				.send(build_400_response(req.body));
+				.send(build_400_response(JSON.stringify(req.body)));
 		}
 
 		let response = await this.service.create(newObject);
@@ -113,10 +78,16 @@ export default class BaseCrudController<T extends Model> extends BaseController 
 	 * @param req request
 	 * @param res response
 	 */
-	async update(req: Request, res: Response) {
-		// @ts-ignore
-		const id = req.params.id;
+	async update(req: IAuthorizedRequest, res: Response) {
+		const id = Number(req.params['id']);
 		const updateData = req.body;
+		let existing = await this.service.getByPk(id);
+		if(!existing.result) {
+			res.status(HttpResponseCode.BAD_REQUEST).send({error: 'Invalid id'});
+		}
+		if (req.user.roleId !== EaasiRoles.ADMIN && existing.result.userId !== req.user.id) {
+			return res.status(HttpResponseCode.UNAUTHORIZED).send();
+		}
 		let updateResponse = await this.service.update(id, updateData);
 
 		if (updateResponse.hasError) {
@@ -131,9 +102,15 @@ export default class BaseCrudController<T extends Model> extends BaseController 
 	 * @param req request
 	 * @param res response
 	 */
-	async delete(req: Request, res: Response) {
-		// @ts-ignore
-		const id = req.params.id;
+	async delete(req: IAuthorizedRequest, res: Response) {
+		const id = Number(req.params['id']);
+		let existing = await this.service.getByPk(id);
+		if(!existing.result) {
+			res.status(HttpResponseCode.BAD_REQUEST).send({error: 'Invalid id'});
+		}
+		if (req.user.roleId !== EaasiRoles.ADMIN && existing.result.userId !== req.user.id) {
+			return res.status(HttpResponseCode.UNAUTHORIZED).send();
+		}
 		let deleteResponse = await this.service.destroy(id);
 
 		if (deleteResponse.hasError) {
@@ -149,7 +126,7 @@ export default class BaseCrudController<T extends Model> extends BaseController 
 	 * @param res response
 	 * @param errors express-validator Result errors
 	 */
-	async sendMalformedRequestResponse(req: Request, res: Response, errors: Result<any>) {
+	async sendMalformedRequestResponse(req: IAuthorizedRequest, res: Response, errors: Result<any>) {
 		let allErrors = errors.array();
 		let errorMessage = '';
 
@@ -177,7 +154,7 @@ export default class BaseCrudController<T extends Model> extends BaseController 
 	 * @returns {Promise<*>}
 	 * @private
 	 */
-	static async _handleUpdateError(req: Request, res: Response, updateResponse: any) {
+	static async _handleUpdateError(req: IAuthorizedRequest, res: Response, updateResponse: any) {
 		if (updateResponse.error === 'notFound') {
 			return res
 				.status(HttpResponseCode.NOT_FOUND)
@@ -197,7 +174,7 @@ export default class BaseCrudController<T extends Model> extends BaseController 
 	 * @returns {Promise<*>}
 	 * @private
 	 */
-	static async _handleDeleteError(req: Request, res: Response, deleteResponse: any) {
+	static async _handleDeleteError(req: IAuthorizedRequest, res: Response, deleteResponse: any) {
 		if (deleteResponse.error === 'notFound') {
 			return res
 				.status(HttpResponseCode.NOT_FOUND)
